Struct glyph_brush::GlyphCalculatorBuilder

pub struct GlyphCalculatorBuilder<'a, H = DefaultSectionHasher> { /* fields omitted */ }

Builder for a GlyphCalculator.

Example

use glyph_brush::GlyphCalculatorBuilder;

let dejavu: &[u8] = include_bytes!("../../fonts/DejaVuSans.ttf");
let mut glyphs = GlyphCalculatorBuilder::using_font_bytes(dejavu).build();

Methods

impl<'a> GlyphCalculatorBuilder<'a>

pub fn using_font_bytes<B: Into<SharedBytes<'a>>>(font_0_data: B) -> Self

Specifies the default font data used to render glyphs. Referenced with FontId(0), which is default.

pub fn using_fonts_bytes<B, V>(font_data: V) -> Self where
B: Into<SharedBytes<'a>>,
V: Into<Vec<B>>,

pub fn using_font(font_0_data: Font<'a>) -> Self

Specifies the default font used to render glyphs. Referenced with FontId(0), which is default.

pub fn using_fonts<V: Into<Vec<Font<'a>>>>(fonts: V) -> Self

impl<'a, H: BuildHasher> GlyphCalculatorBuilder<'a, H>

pub fn add_font_bytes<B: Into<SharedBytes<'a>>>(
&mut self,
font_data: B
) -> FontId

Adds additional fonts to the one added in using_font / using_font_bytes.

Returns a FontId to reference this font.

pub fn add_font(&mut self, font_data: Font<'a>) -> FontId

Adds additional fonts to the one added in using_font / using_font_bytes.

Returns a FontId to reference this font.

pub fn section_hasher<T: BuildHasher>(
self,
section_hasher: T
) -> GlyphCalculatorBuilder<'a, T>

Sets the section hasher. GlyphCalculator cannot handle absolute section hash collisions so use a good hash algorithm.

This hasher is used to distinguish sections, rather than for hashmap internal use.

Defaults to xxHash.

pub fn build(self) -> GlyphCalculator<'a, H>

Builds a GlyphCalculator